Use internal iteration over hash tables in Aggregator.
For compound hash tables such as the future StringHashMap, implementing
an iterator-based API might be inefficient, because:
1) the information about the component to which the value belongs is
lost, and has to be recalculated if the value is reinserted;
2) the key has to be converted to a general format from a from a
component-specific format, which may differ between the components.
These inefficiencies are especially apparent when merging two hash
tables. In contrast, an efficient algorithm would merge a compound table
component-wise, avoiding unnecessary conversions and recalculations.

Therefore, switch Aggregator to use internal iteration (map-like
functions) when traversing and merging the hash tables.

This commit prepares for the StringHashMap PR #5417 and contains some of
its code.